backend/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
pydantic==2.11.9
python-dotenv==1.1.1
sqlalchemy==2.0.43
alembic==1.16.5
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-jose[cryptography]==3.5.0
